Replace formatValueShort with shared formatValue utility
Create format-value.ts with formatValue(value, 'inline-hint') for
non-React contexts like Monaco decorations. Delete formatValueShort
from log-decorations.ts. Both proto-traversal format functions are
now eliminated.
